+++ src/main/java/com/spring_boot_final/project/controller/exhbn/ExhbnViewController.java
@@ -37,22 +37,5 @@ public class ExhbnViewController {
         return "list";
     }
 
-    // course page view
-    @RequestMapping("/course")
-    public String courseView(Model model) {
 
-        ArrayList<ExhbnVO> vo = service.selectAllData();
-        model.addAttribute("exhbnList",vo);
-
-        return "course/courseMain";
-
-    // course board view
-    }@RequestMapping("/course/board")
-    public String courseBoardView(Model model) {
-
-        ArrayList<ExhbnVO> vo = service.selectAllData();
-        model.addAttribute("exhbnList",vo);
-
-        return "course/courseBoard";
-    }
 }
